requirements.txt
python>=3.13

# Database
psycopg2-binary==2.9.9

python-dotenv==1.1.0

pandas==2.3

transformers==4.51
